Update 숫자 카드 2.cc

+++ 백준/Silver/10816. 숫자 카드 2/숫자 카드 2.cc
@@ -7,10 +7,10 @@ int main(){
     cin.tie(NULL);
     cout.tie(NULL);
     int N,M;
-    map<long long,int> cards;
+    map<long,int> cards;
     cin>>N;
     for(int i = 0;i<N;i++){
-        long long input;
+        long input;
         cin>>input;
         if(cards.find(input) != cards.end())
             cards[input]+=1;
@@ -19,7 +19,7 @@ int main(){
     }
     cin>>M;
     for(int i = 0;i<M;i++){
-        long long input;
+        long input;
         cin>>input;
         if(cards.find(input) != cards.end())
             cout<<cards[input]<<' ';
